Homologação/Streamlit/RH Formulário/requirements.txt
pytest==7.2.1
pandas==1.5.3
streamlit==1.22.0
PyPDF2==3.0.1
python-docx==0.8.11
openpyxl==3.1.0
streamlit-js-eval==0.1.5